Translation to spanish
The nex translation take es_VE and make it panhispanis, so now es is
used for all the spanish variants.

LINGUAS was updated to reflect the change.

aditionally using the built-in capabilities of poedit I pulled the
strings with the translatable thing from the source and updated the
translation fo the 100%, and recheck it with some styleguides from
some videogame companies and online forums to chech for consistencies
in terminology.

the remained untranslated parts are non translatable (for now) on the
source code, I will check if I can find them and update.

Signed-off-by: Pablo Lezaeta Reyes <prflr88@gmail.com>